chore: move route based traversal methods and scopes into separate files in preparation of a new traversal implementation
