+++ app/models/namespace.rb
@@ -6,6 +6,7 @@ class Namespace < ApplicationRecord # rubocop:disable Metrics/ClassLength
   include HasPuid
   include Routable
   include TrackActivity
+  include Namespaces::Traversal::RouteBased
 
   has_logidze
   acts_as_paranoid
@@ -70,10 +71,6 @@ class Namespace < ApplicationRecord # rubocop:disable Metrics/ClassLength
       find_by(arel_table[:path].lower.eq(path.downcase))
     end
 
-    def as_ids
-      select(Namespace.arel_table[:id])
-    end
-
     # Return common group ancestor IDs between two namespace parents.
     # This can be reused by operations that need to skip shared ancestors.
     # @param old_parent [Namespace, nil] the previous parent namespace
@@ -100,104 +97,6 @@ class Namespace < ApplicationRecord # rubocop:disable Metrics/ClassLength
       old_parent&.propagate_samples_count_delta(-sample_count, except_group_ancestor_ids: common_group_ids)
       # Add to new parent ancestors except common ancestors.
       new_parent.propagate_samples_count_delta(sample_count, except_group_ancestor_ids: common_group_ids)
-    end
-
-    # Return a relation containing this namespace (or namespaces from the
-    # receiver relation) and all ancestor namespaces.
-    #
-    # Implementation:
-    # - Builds an Arel subquery that finds route ids for any route that is an
-    #   ancestor of the current namespace by matching path prefixes. The
-    #   subquery uses a joined (aliased) `ancestral_routes` table and matches
-    #   either the same id or a path prefix using `CONCAT(path, '/') LIKE
-    #   CONCAT(ancestral_routes.path, '/%')`.
-    # - Filters to routes whose `source_type` is the Namespace STI class and
-    #   whose `source_id` is within the set of namespaces in the original
-    #   relation (`select(:id).arel`). The result is a distinct list of route
-    #   ids which is then used to find matching Namespace records via a
-    #   join on `:route`.
-    #
-    # Returns an ActiveRecord::Relation of Namespace records (self + ancestors).
-    def self_and_ancestors # rubocop:disable Metrics/AbcSize
-      # build sql expression to select the route ids of the self and ancestral groups
-      ancestral_routes = Arel::Table.new(Route.table_name, as: 'ancestral_routes')
-      ancestral_route_ids = Route.arel_table.join(ancestral_routes, Arel::Nodes::OuterJoin).on(
-        ancestral_routes[:id].eq(Route.arel_table[:id]).or(
-          concat_path_with_slash(Route.arel_table[:path]).matches(
-            concat_path_with_wildcard(ancestral_routes[:path])
-          )
-        )
-      ).where(
-        Route.arel_table[:source_type].eq(Namespace.sti_name).and(
-          Route.arel_table[:source_id].in(select(:id).arel)
-        ).and(Route.arel_table[:deleted_at].eq(nil))
-      ).project(Route.arel_table[:id]).distinct
-
-      unscoped
-        .joins(:route)
-        .where(Route.arel_table[:id].in(ancestral_route_ids))
-    end
-
-    # Return a relation containing this namespace and all descendant
-    # namespaces.
-    #
-    # Implementation:
-    # - Similar approach to `self_and_ancestors` but matches descendant paths
-    #   by checking whether the descendant route's path matches the current
-    #   namespace's path with a trailing '/%'. Uses an aliased
-    #   `descendant_routes` table and constructs a distinct list of matching
-    #   route ids which is used to filter namespaces.
-    #
-    # Returns an ActiveRecord::Relation of Namespace records (self + descendants).
-    def self_and_descendants # rubocop:disable Metrics/AbcSize
-      # build sql expression to select the route ids of the self and descendant groups
-      descendant_routes = Arel::Table.new(Route.table_name, as: 'descendant_routes')
-      descendant_route_ids = Route.arel_table.join(descendant_routes, Arel::Nodes::OuterJoin).on(
-        descendant_routes[:id].eq(Route.arel_table[:id]).or(
-          descendant_routes[:path].matches(
-            concat_path_with_wildcard(Route.arel_table[:path])
-          )
-        )
-      ).where(
-        Route.arel_table[:source_type].eq(Namespace.sti_name).and(
-          Route.arel_table[:source_id].in(select(:id).arel)
-        ).and(Route.arel_table[:deleted_at].eq(nil))
-      ).project(descendant_routes[:id]).distinct
-
-      unscoped
-        .joins(:route)
-        .where(Route.arel_table[:id].in(descendant_route_ids))
-    end
-
-    def self_and_descendant_ids
-      self_and_descendants.as_ids
-    end
-
-    # Return namespaces that do not have descendants inside the provided
-    # collection. This is useful when you want to filter a set of namespaces
-    # to only the top-level ones (no child namespaces included).
-    #
-    # Implementation:
-    # - Builds a subquery producing wildcard paths (CONCAT(path, '/%')) for
-    #   each joined route and then filters namespaces whose path does not
-    #   match ANY of those wildcard patterns using PostgreSQL's `NOT LIKE ALL`
-    #   semantics (`does_not_match` wraps `NOT ILIKE`/`NOT LIKE` depending on
-    #   adapter).
-    #
-    # Returns an ActiveRecord::Relation of Namespace records that have no
-    # descendants within the given relation.
-    def without_descendants
-      wildcard_path_select =
-        joins(:route)
-        .select(concat_path_with_wildcard(Route.arel_table[:path])).arel
-
-      joins(:route)
-        .where(Route.arel_table[:path].does_not_match(
-                 Arel::Nodes::NamedFunction.new('ALL',
-                                                [Arel::Nodes::NamedFunction.new(
-                                                  'ARRAY', [wildcard_path_select]
-                                                )])
-               ))
     end
 
     def ransackable_attributes(_auth_object = nil)
@@ -435,92 +334,6 @@ class Namespace < ApplicationRecord # rubocop:disable Metrics/ClassLength
         ), Arel::Nodes::SqlLiteral.new('jsonb')
       )
     end
-  end
-
-  def ancestors
-    return self.class.none if parent_id.blank?
-
-    self_and_ancestors.where.not(id:)
-  end
-
-  def ancestor_ids
-    ancestors.as_ids
-  end
-
-  def self_and_ancestors
-    return self.class.where(id:) if parent_id.blank?
-
-    self_and_ancestors_of_type(self.class.sti_name)
-  end
-
-  # Return namespaces whose routes are ancestors of the current namespace
-  # (filtered by the provided STI `types`). This variant accepts an explicit
-  # set of types and is used by `self_and_ancestors` to restrict results to
-  # particular namespace subtypes (e.g. Group, ProjectNamespace).
-  #
-  # Implementation details:
-  # - Projects a CONCAT(path, '/') expression from the `routes` table for the
-  #   current set of routes and matches it against routes whose path begins
-  #   with that projected value using a trailing wildcard (CONCAT(path, '/%')).
-  # - Uses Arel functions to build the CONCAT and LIKE expressions because
-  #   the SQL is Postgres-specific and not modelled directly by Arel.
-  # - We call the helpers on `self.class` because `concat_path_with_slash` and
-  #   `concat_path_with_wildcard` are defined as class-level helpers inside the
-  #   `class << self` block above. Calling them via `self.class` keeps this
-  #   instance-level method concise while reusing the shared Arel builders.
-  #
-  # @param types [Array<String>, String] STI type(s) to filter the namespaces
-  # @return [ActiveRecord::Relation<Namespace>] matching ancestor namespaces of the given types
-  def self_and_ancestors_of_type(types)
-    Namespace.joins(:route)
-             .where(
-               Arel::Nodes::Grouping.new(
-                 Route.arel_table.project(
-                   self.class.concat_path_with_slash(Route.arel_table[:path])
-                 ).where(Route.arel_table[:source_id].eq(id))
-               ).matches(self.class.concat_path_with_wildcard(Route.arel_table[:path]))
-             ).where(type: types)
-  end
-
-  def self_and_ancestor_ids
-    self_and_ancestors.as_ids
-  end
-
-  def descendants
-    self_and_descendants.where.not(id:)
-  end
-
-  def self_and_descendants
-    self_and_descendants_of_type(self.class.sti_name)
-  end
-
-  # Return namespaces whose routes are descendants of the current namespace
-  # (filtered by the provided STI `types`). This variant accepts an explicit
-  # set of types and is used by `self_and_descendants` to restrict results to
-  # particular namespace subtypes (e.g. Group, ProjectNamespace).
-  #
-  # Implementation details:
-  # - Matches routes whose `path` begins with the current namespace's
-  #   `full_path` (or `full_path/` followed by anything) using SQL `LIKE` with
-  #   a trailing '/%' wildcard. We use Arel's `matches_any` for the OR-style
-  #   pattern matching against both exact path and prefixed descendants.
-  # - This method keeps the Arel expression straightforward by delegating
-  #   the path concatenation and wildcard creation to the class-level helpers
-  #   where appropriate; here we inline a `matches_any` over the two patterns
-  #   since the descendants check is simple and avoids extra subqueries.
-  #
-  # @param types [Array<String>, String] STI type(s) to filter the namespaces
-  # @return [ActiveRecord::Relation<Namespace>] matching descendant namespaces of the given types
-  def self_and_descendants_of_type(types)
-    route_path = Route.arel_table[:path]
-
-    Namespace.joins(:route)
-             .where(route_path.matches_any([full_path, "#{full_path}/%"]))
-             .where(type: types)
-  end
-
-  def self_and_descendant_ids
-    self_and_descendants.as_ids
   end
 
   def to_param
